Rename iter module to backend.
